refactor: whole agent scope logic
